+++ matrix-sdk/src/main/java/org/matrix/androidsdk/data/store/MXMemoryStore.java
@@ -1119,7 +1119,7 @@ public class MXMemoryStore implements IMXStore {
                 for (int index = events.size() - 1; index >= 0; index--) {
                     Event event = eventsList.get(index);
 
-                    if (event.isUnkownDevice()) {
+                    if (event.isUnknownDevice()) {
                         unknownDeviceEvents.add(event);
                     }
                 }