+++ CHANGES.rst
@@ -12,6 +12,7 @@ Bugfix:
 
 API Change:
  - New API: add device_id param to LoginRestClient.loginWithUser()
+ - API change: Event.isUnkownDevice() as been renamed to Event.isUnknownDevice() (typo)
 
 Translations:
  -